Dedup for all & added long-term compressed files.

+++ sh/ghGenerate.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 COMPRESS_CRIT="\.(ass|atom|bin|bm|bmp|conf|css|csv|htm|html|ico|js|json|kar|list|lrc|lst|map|md|mid|mjs|mod|mts|otf|rss|sbv|srt|ssa|svg|trc|ts|tsv|ttf|ttml|txt|vgm|vtt|wasm|webmanifest|xml|ytt)$"
 
-#if [ ! -f "$(which zopfli)" ]; then
-	#sudo apt install -y zopfli
-#fi
-if [ ! -f "$(which tree)" ]; then
-	sudo apt install -y tree
+if [ "$1" != "" ]; then
+	if [ ! -f "$(which zopfli)" ]; then
+		sudo apt install -y zopfli
+	fi
 fi
+#if [ ! -f "$(which tree)" ]; then
+	#sudo apt install -y tree
+#fi
 
 echo "$(date +"%s")" > build-time.txt
 cp -Lr ghp ghp-raw
@@ -42,9 +44,34 @@ tree -ifl | while IFS= read -r file; do
 	if [ -f "$file" ]; then
 		# Is a file
 		if [ "$(echo "$file" | grep -E "$COMPRESS_CRIT")" != "" ]; then
-			rm "$file"
+			if [[ "$file" == *"/index.htm" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			elif [[ "$file" == *"/index.html" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			elif [[ "$file" == *"/default.htm" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			elif [[ "$file" == *"/default.html" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			else
+				rm "$file"
+			fi
 		else
-			echo "File \"${file}\" is preserved."
+			fileHash="$(sha256sum "${file}" | cut -d' ' -f1)"
+			findResult="$(grep -F "${fileHash}	" ../fileHashes.tsv | cut -d '	' -f2)"
+			if [ "$findResult" != "" ] ; then
+				pathDiffRaw="$(realpath -Lsm --relative-to="${file}" "${findResult}")"
+				pathDiff="${pathDiffRaw/\.\.\//}"
+				echo "Deduplicated: ${file} -> ${pathDiff}"
+				rm "${file}"
+				ln -s "${pathDiff}" "${file}"
+			else
+				echo "${fileHash}	$(realpath -s "${file}")" >> ../fileHashes.tsv
+				echo "File \"${file}\" is preserved."
+			fi
 		fi
 	fi
 done
@@ -74,7 +101,11 @@ tree -ifl | while IFS= read -r file; do
 				rm "$file"
 			else
 				echo "${fileHash}	$(realpath -s "${file}")" >> ../fileHashes.tsv
-				gzip -9 "$file" && echo "Compressed \"${file}\" with Gzip."
+				if [ "$1" != "" ]; then
+					zopfli --i1 && echo "Compressed \"${file}\" with Zopfli."
+				else
+					gzip -9 "$file" && echo "Compressed \"${file}\" with Gzip."
+				fi
 			fi
 		#else
 			#echo "File \"${file}\" cannot be compressed."
@@ -111,7 +142,11 @@ tree -ifl | while IFS= read -r file; do
 				rm "$file"
 			else
 				echo "${fileHash}	$(realpath -s "${file}")" >> ../fileHashes.tsv
-				brotli -v9j "$file"
+				if [ "$1" != "" ]; then
+					brotli -vjq 10 "$file"
+				else
+					brotli -v9j "$file"
+				fi
 			fi
 		#else
 			#echo "File \"${file}\" cannot be compressed."
